+++ examples/webpack-esm/cypress.config.mjs
@@ -3,6 +3,7 @@ import webpack from "@cypress/webpack-preprocessor";
 import preprocessor from "@badeball/cypress-cucumber-preprocessor";
 
 export async function setupNodeEvents(on, config) {
+  // This is required for the preprocessor to be able to generate JSON reports after each run, and more,
   await preprocessor.addCucumberPreprocessorPlugin(on, config);
 
   on(